SEO and URL handling logic reorganization:
