--- src/VirtoCommerce.News.Core/Services/INewsArticleSeoResolver.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using VirtoCommerce.Seo.Core.Models;
-using VirtoCommerce.Seo.Core.Services;
-
-namespace VirtoCommerce.News.Core.Services;
-
-public interface INewsArticleSeoResolver : ISeoResolver
-{
-    public Task<IList<SeoInfo>> FindActiveSeoAsync(string[] linkSegments, string storeId, string languageCode);
-}
