+++ src/VirtoCommerce.News.Core/Services/INewsArticleSeoService.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using VirtoCommerce.Seo.Core.Models;
+
+namespace VirtoCommerce.News.Core.Services;
+
+public interface INewsArticleSeoService
+{
+    Task<IList<SeoInfo>> FindActiveSeoAsync(string[] linkSegments, string storeId, string languageCode);
+}
