+++ src/VirtoCommerce.News.ExperienceApi/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,7 @@
 using GraphQL.MicrosoftDI;
 using Microsoft.Extensions.DependencyInjection;
+using VirtoCommerce.News.ExperienceApi.Services;
+using VirtoCommerce.Seo.Core.Services;
 using VirtoCommerce.Xapi.Core.Extensions;
 using VirtoCommerce.Xapi.Core.Infrastructure;
 
@@ -15,6 +17,11 @@ public static class ServiceCollectionExtensions
         });
 
         serviceCollection.AddSingleton<ScopedSchemaFactory<XapiAssemblyMarker>>();
+
+        serviceCollection.AddTransient<INewsArticleSettingsService, NewsArticleSettingsService>();
+
+        //SEO
+        serviceCollection.AddTransient<ISeoResolver, NewsArticleSeoResolver>();
 
         return serviceCollection;
     }
