+++ src/VirtoCommerce.News.ExperienceApi/Services/INewsArticleSettingsService.cs
@@ -1,0 +1,8 @@
+using System.Threading.Tasks;
+
+namespace VirtoCommerce.News.ExperienceApi.Services;
+
+public interface INewsArticleSettingsService
+{
+    Task<bool> GetUseRootLinkSettingAsync();
+}
